+++ src/main/java/com/br/cooapi/combustivel/CombustivelRepositories.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CombustivelRepositories extends JpaRepository<Combustivel,Integer> {
+public interface CombustivelRepositories extends JpaRepository<Combustivel, Long> {
 }
